Splitted extensions files into several files.
Issue #97
