+++ GiniVision/Classes/Extensions/AVCaptureVideoOrientation.swift
@@ -1,0 +1,32 @@
+//
+//  AVCaptureVideoOrientation.swift
+//  GiniVision
+//
+//  Created by Enrique del Pozo Gómez on 10/30/17.
+//  Copyright © 2017 Gini GmbH. All rights reserved.
+//
+
+import AVFoundation
+
+internal extension AVCaptureVideoOrientation {
+    
+    internal init(_ interface: UIInterfaceOrientation) {
+        switch interface {
+        case .portraitUpsideDown: self = .portraitUpsideDown
+        case .landscapeLeft: self = .landscapeLeft
+        case .landscapeRight: self = .landscapeRight
+        default: self = .portrait
+        }
+    }
+    
+    internal init?(_ device: UIDeviceOrientation?) {
+        guard let orientation = device else { return nil }
+        switch orientation {
+        case .portrait: self = .portrait
+        case .portraitUpsideDown: self = .portraitUpsideDown
+        case .landscapeLeft: self = .landscapeRight
+        case .landscapeRight: self = .landscapeLeft
+        default: return nil
+        }
+    }
+}
