+++ GiniVision/Classes/Extensions/Sequence.swift
@@ -1,0 +1,17 @@
+//
+//  Sequence.swift
+//  GiniVision
+//
+//  Created by Enrique del Pozo Gómez on 10/30/17.
+//  Copyright © 2017 Gini GmbH. All rights reserved.
+//
+
+import Foundation
+
+internal extension Sequence {
+    func shuffled() -> [Iterator.Element] {
+        var result = Array(self)
+        result.shuffle()
+        return result
+    }
+}
